In malloc(Tau_memorywrap_add_ptr), we were using "size_t" for size, but in free (Tau_memorywrap_remove_ptr), we were using "int". This probably gave us the negative numbers for large sizes (> 4gb). Changed int to size_t.

+++ src/Profile/TauMemoryWrap.cpp
@@ -214,7 +214,7 @@ extern "C" void Tau_memorywrap_remove_ptr (void *ptr) {
     RtsLayer::LockDB();
     TAU_HASH_MAP<void*,MemoryAllocation>::const_iterator it = global().pointerMap.find(ptr);
     if (it != global().pointerMap.end()) {
-      int size = global().pointerMap[ptr].numBytes;
+      size_t size = global().pointerMap[ptr].numBytes;
       global().pointerMap.erase(ptr);
       if (size > 0) {
         global().bytesAllocated -= size;
